Fix query issue when deleting unused categories

+++ app/Http/Controllers/MonthlyTrackingController.php
@@ -141,7 +141,7 @@ class MonthlyTrackingController extends Controller
       $request->session()->put('saved', $record->occurred_at);
 
       if($oldCategory && $oldCategory != $record->category){
-        $this->checkCategory(Auth::user()->id, $record->month_id, $oldCategory);
+        $this->checkCategory(Auth::user()->id, $oldCategory);
       }
 
       return Redirect::back();
@@ -227,19 +227,24 @@ class MonthlyTrackingController extends Controller
       return $monthlyBudgetCategories->pluck('description')->toJson();
     }
 
-    function checkCategory($user, $trackedMonth, $name)
+    function checkCategory($user, $name)
     {
-      $categoryRecords = MonthlyTrackingRecord::where('month_id', $trackedMonth)->where('category', $name)->get();
+      $categoryRecords = MonthlyTrackingRecord::whereHas('trackedMonth', function($query) use($user) {
+                                                  $query->where('user_id', $user);
+                                                })->where('category', $name)->get();
       if($categoryRecords->isEmpty()){
         //the category is no longer being used. Delete it.
         $budgetCategory = MonthlyBudgetRecord::where('user_id', $user)->where('description', $name)->first();
 
-        $categoryValues =  MonthlyBudgetRecordValue::where('record_id', $budgetCategory->id)->get();
 
-        foreach ($categoryValues as $value) {
-          $value->delete();
-        }
-        $budgetCategory->delete();
+        if($budgetCategory){
+          $categoryValues =  MonthlyBudgetRecordValue::where('record_id', $budgetCategory->id)->get();
+
+          foreach ($categoryValues as $value) {
+            $value->delete();
+          }
+          $budgetCategory->delete();
+         }
       }
     }
